Add code for esp8266 initial setup
Added code (remarked out) that will only be needed once to get esp8266 to auto-connect to your home wifi.  Note: I did not use the Arduino to send this code (like the script implies).  I reverse connected the esp8266 (with logic level converters) to the Arduino board (with the Arduino chip forced into reset) and used the usb to send commands directly to the esp8266.